% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cnes.R
\name{get_cnes}
\alias{get_cnes}
\title{Get health units data from CNES}
\usage{
get_cnes(
  conn,
  agr,
  ano_competen = NULL,
  mes_competen = NULL,
  vinc_sus = NULL,
  leitos_hosp = NULL,
  tipo_unidade = NULL,
  esfera_admin = NULL,
  lucene_query = NULL
)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano_competen}{numeric. Competence year.}

\item{mes_competen}{numeric. Competence month.}

\item{vinc_sus}{numeric. Health units with SUS relationship. 1 for yes and 0 for no.}

\item{leitos_hosp}{numeric. Health units with hospital, circury and other kinds of beds.}

\item{tipo_unidade}{string. Health unit type.}

\item{esfera_admin}{string. Health unit administration level.}

\item{lucene_query}{string. A more complex lucene query can be presented here. The string must be informed using sigle quotes. If used, all other filters will be ignored.}

\item{agg}{string. Aggregation level. 'mun' for municipalities, 'uf' for "unidades federativas" or 'regsaude' for "regiões de saúde".}
}
\value{
A \code{data.frame} containing number of hospital admission authorizations (\code{sih}) for the aggregation level.
}
\description{
This function retrieves health units data from PCDaS ElasticSearch cluster for a specified year.
}
\details{
This function uses raw data from the Cadastro Nacional de Estabelecimentos de Saúde Sistema de Informações de Internações Hospitalares (SIH) available at the PCDaS ElasticSearch cluster. A documentation about this data can be found at \url{https://bigdata-metadados.icict.fiocruz.br/dataset/sistema-de-informacoes-de-mortalidade-sim}.
The CID-10 codes and name for chapters, group, category and subcategories of basic cause of death can be inspected at the same address above.
}
\examples{
sim <- get_sih(conn = conn, ano = 2010, agr = "mun", sexo = "Masculino")
sim <- get_sih(conn = conn, ano = 2010, agr = "mun", diag_princ = "J448")
sim <- get_sih(conn = conn, ano = 2010, agr = "mun", idade_anos_min = 0, idade_anos_max = 10)
sim <- get_sih(conn = conn, ano = 2010, agr = "mun", proced_real = 310010039)
sih <- get_sih(conn = conn, agr = "mun", lucene_query = 'ano_internacao: 2010 AND def_raca_cor: "Preta" or "Branca"')
}
