% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_age.R
\name{get_pop_age}
\alias{get_pop_age}
\title{Get population from SVS with sex and age}
\usage{
get_pop_age(conn, ano, agr)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year.}

\item{agg}{string. Aggregation level. 'mun' for municipalities, 'uf' for "unidades federativas" or 'regsaude' for "regiões de saúde".}
}
\value{
A \code{data.frame} containing population for the aggregation level.
}
\description{
This function retrieves population data from PCDaS ElasticSearch cluster for a specified year.
}
\examples{
pop_df <- get_pop_mun(conn = conn, ano = 2010, agr = "mun")
}
