% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_mun.R
\name{get_pop_mun}
\alias{get_pop_mun}
\title{Get population for municipalities}
\usage{
get_pop_mun(conn, ano)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year.}
}
\value{
A \code{data.frame} containing the municipalities IBGE codes (\code{cod_mun}) and population (\code{pop}).
}
\description{
This function retrieves population for all Brazilian municipalities from PCDaS ElasticSearch cluster for a specified year.
}
\examples{
pop_df <- get_pop_mun(conn = conn, ano = 2010)
}
