% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sih.R
\name{get_sih}
\alias{get_sih}
\title{Get mortality data from SIH}
\usage{
get_sih(
  conn,
  agr,
  ano = NULL,
  sexo = NULL,
  diag_princ = NULL,
  diag_princ_capitulo = NULL,
  diag_princ_grupo = NULL,
  diag_princ_categoria = NULL,
  diag_princ_subcategoria = NULL,
  idade_anos_min = NULL,
  idade_anos_max = NULL,
  proced_real = NULL,
  lucene_query = NULL
)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year of death.}

\item{sexo}{string. Sex of the patient \code{Masculino} for males or \code{Feminino} for females.}

\item{diag_princ}{string. CID-10 code for the main diagnosis.}

\item{diag_princ_capitulo}{string. Chapter of the main diagnosis.}

\item{diag_princ_grupo}{string. Group of the main diagnosis.}

\item{diag_princ_categoria}{string. Category of the main diagnosis.}

\item{diag_princ_subcategoria}{string. Subcategory of the main diagnosis.}

\item{idade_anos_min}{numeric. Minimum age of patient, in years.}

\item{idade_anos_max}{numeric. Maximum age of patient, in years.}

\item{proced_real}{numeric. Numeric code of the realized hospital procedure.}

\item{lucene_query}{string. A more complex lucene query can be presented here. The string must be informed using sigle quotes. If used, all other filters will be ignored.}

\item{agg}{string. Aggregation level. 'mun' for municipalities, 'uf' for "unidades federativas" or 'regsaude' for "regiões de saúde".}
}
\value{
A \code{data.frame} containing number of hospital admission authorizations (\code{sih}) for the aggregation level.
}
\description{
This function retrieves hospital admission data from PCDaS ElasticSearch cluster for a specified year.
}
\details{
This function uses raw data from the Sistema de Informações de Internações Hospitalares (SIH) available at the PCDaS ElasticSearch cluster. A documentation about this data can be found at \url{https://bigdata-metadados.icict.fiocruz.br/dataset/sistema-de-informacoes-de-mortalidade-sim}.
The CID-10 codes and name for chapters, group, category and subcategories of basic cause of death can be inspected at the same address above.
}
\examples{
sim <- get_sih(conn = conn, ano = 2010, agr = "mun", sexo = "Masculino")
sim <- get_sih(conn = conn, ano = 2010, agr = "mun", diag_princ = "J448")
sim <- get_sih(conn = conn, ano = 2010, agr = "mun", idade_anos_min = 0, idade_anos_max = 10)
sim <- get_sih(conn = conn, ano = 2010, agr = "mun", proced_real = 310010039)
sih <- get_sih(conn = conn, agr = "mun", lucene_query = 'ano_internacao: 2010 AND def_raca_cor: "Preta" or "Branca"')
}
