% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim.R
\name{get_sim}
\alias{get_sim}
\title{Get mortality data from SIM}
\usage{
get_sim(conn, ano, agr, sexo = NULL, causabas = NULL,
  causabas_capitulo = NULL, causabas_grupo = NULL,
  causabas_categoria = NULL, causabas_subcategoria = NULL,
  idade_obito_anos_min = NULL, idade_obito_anos_max = NULL,
  idade_obito_dias_min = NULL, idade_obito_dias_max = NULL)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year of death.}

\item{sexo}{string. Sex of the deceased. \code{Masculino} for males or \code{Feminino} for females.}

\item{causabas}{string. CID-10 code for the basic cause of death.}

\item{causabas_capitulo}{string. Chapter of the basic cause of death.}

\item{causabas_grupo}{string. Group of the basic cause of death.}

\item{causabas_categoria}{string. Category of the basic cause of death.}

\item{causabas_subcategoria}{string. Subcategory of the basic cause of death.}

\item{idade_obito_anos_min}{numeric. Minimum age of death, in years.}

\item{idade_obito_anos_max}{numeric. Maximum age of death, in years.}

\item{idade_obito_dias_min}{numeric. Minimum age of death, in days.}

\item{idade_obito_dias_max}{numeric. Maximum age of death, in days.}

\item{agg}{string. Aggregation level. 'mun' for municipalities, 'uf' for "unidades federativas" or 'regsaude' for "regiões de saúde".}
}
\value{
A \code{data.frame} containing number of deceased (\code{sim}) for the aggregation level.
}
\description{
This function retrieves mortality data from PCDaS ElasticSearch cluster for a specified year.
}
\details{
This function uses raw data from the Sistema de Informações de Mortalidade (SIM) available at the PCDaS ElasticSearch cluster. A documentation about this data can be found at \url{https://bigdata-metadados.icict.fiocruz.br/dataset/sistema-de-informacoes-de-mortalidade-sim}.
The CID-10 codes and name for chapters, group, category and subcategories of basic cause of death can be inspected at the same address above.
}
\examples{
sim <- get_sim(conn = conn, ano = 2010, agr = "mun", sexo = "Masculino")
sim <- get_sim(conn = conn, ano = 2010, agr = "mun", causabas = "R98")
sim <- get_sim(conn = conn, ano = 2010, agr = "mun", causabas_categoria = "O01   Mola hidatiforme")
sim <- get_sim(conn = conn, ano = 2010, agr = "mun", idade_obito_anos_min = 0, idade_obito_anos_max = 1)
sim <- get_sim(conn = conn, ano = 2010, agr = "mun", idade_obito_dias_min = 0, idade_obito_dias_max = 1)
}
