% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_capitulos.R
\name{get_sim_capitulos}
\alias{get_sim_capitulos}
\title{Get mortality data from SIM for several capítulos}
\usage{
get_sim_capitulos(conn, ano, agr, causabas_capitulos)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year of death.}

\item{causabas_capitulos}{string vector of desired capítulos.}

\item{agg}{string. Aggregation level. 'mun' for municipalities, 'uf' for "unidades federativas" or 'regsaude' for "regiões de saúde".}
}
\value{
A \code{data.frame} containing number of deceased (\code{sim}) for the aggregation level.
}
\description{
This function retrieves mortality data from PCDaS ElasticSearch cluster for a specified year and for several capítulos.
}
\details{
This function uses raw data from the Sistema de Informações de Mortalidade (SIM) available at the PCDaS ElasticSearch cluster. A documentation about this data can be found at \url{https://bigdata-metadados.icict.fiocruz.br/dataset/sistema-de-informacoes-de-mortalidade-sim}.
The CID-10 codes and name for chapters, group, category and subcategories of basic cause of death can be inspected at the same address above.
}
\examples{
sim <- get_sim_capitulos(conn = conn, ano = 2010, agr = "mun", causabas_capitulos = c("I.   Algumas doenças infecciosas e parasitárias", "II.  Neoplasias (tumores)"))
}
