% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sinasc_mun.R
\name{get_sinasc_mun}
\alias{get_sinasc_mun}
\title{Get live births data for municipalities}
\usage{
get_sinasc_mun(conn, ano)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year of birth.}
}
\value{
A \code{data.frame} containing the municipalities IBGE codes (\code{cod_mun}) and number of live births (\code{sinasc}).
}
\description{
This function retrieves live births data for all Brazilian municipalities from PCDaS ElasticSearch cluster for a specified year.
}
\details{
This function uses raw data from the Sistema de Informações de Nascidos Vivos (SINASC) available at the PCDaS ElasticSearch cluster. A documentation about this data can be found at \url{https://bigdata-metadados.icict.fiocruz.br/dataset/sistema-de-informacoes-de-nascidos-vivos-sinasc}.
}
\examples{
sinasc <- get_sinasc_mun(conn = conn, ano = 2010)
}
