% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi_c.1.3.R
\name{indi_c.1.3}
\alias{indi_c.1.3}
\title{Taxa de mortalidade pós-neonatal - C.1.3}
\usage{
indi_c.1.3(conn, ano, multi = 1000)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year.}

\item{multi}{Indicator multiplier. Defaults to RIPSA recommendation.}
}
\value{
A \code{data.frame} containing the municipalities IBGE codes (\code{cod_mun}) and the calculated indicator.
}
\description{
Follows the RIPSA 2012 card \url{http://fichas.ripsa.org.br/2012/c-1-3}
}
\examples{
c.1.3 <- indi_c.1.3(conn = conn, ano = 2010)
}
