% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi_c.10.R
\name{indi_c.10}
\alias{indi_c.10}
\title{Taxa de mortalidade específica por neoplasias malignas - C.10}
\usage{
indi_c.10(conn, ano, agr, multi = 1e+05)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year.}

\item{agr}{string. Aggregation level. 'mun' for municipalities, 'uf' for "unidades federativas" or 'regsaude' for "regiões de saúde".}

\item{multi}{Indicator multiplier. Defaults to RIPSA recommendation.}
}
\value{
A \code{data.frame} containing the calculated indicator for the aggregation level.
}
\description{
Follows the RIPSA 2012 card \url{http://fichas.ripsa.org.br/2012/c-10}
}
\examples{
c.10 <- indi_c.10(conn, 2010, "mun")
}
