% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi_c.12.R
\name{indi_c.12}
\alias{indi_c.12}
\title{Taxa de mortalidade específica por diabete melito - C.12}
\usage{
indi_c.12(conn, ano, multi = 1e+05)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year.}

\item{multi}{Indicator multiplier. Defaults to RIPSA recommendation.}
}
\value{
A \code{data.frame} containing the municipalities IBGE codes (\code{cod_mun}) and the calculated indicator.
}
\description{
Follows the RIPSA 2012 card \url{http://fichas.ripsa.org.br/2012/c-12}
}
\examples{
c.12 <- indi_c.12(conn, 2010)
}
