% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi_c.15.R
\name{indi_c.15}
\alias{indi_c.15}
\title{Taxa de mortalidade específica por afecções originadas no período perinatal - C.15}
\usage{
indi_c.15(conn, ano, multi = 1000)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year.}

\item{multi}{Indicator multiplier. Defaults to RIPSA recommendation.}
}
\value{
A \code{data.frame} containing the municipalities IBGE codes (\code{cod_mun}) and the calculated indicator.
}
\description{
Follows the RIPSA 2012 card \url{http://fichas.ripsa.org.br/2012/c-15}
}
\examples{
c.15 <- indi_c.15(conn, 2010)
}
