% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi_c.4.R
\name{indi_c.4}
\alias{indi_c.4}
\title{Mortalidade proporcional por grupos de causas - C.4}
\usage{
indi_c.4(conn, ano, agr, multi = 100)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year.}

\item{agr}{string. Aggregation level. 'mun' for municipalities, 'uf' for "unidades federativas" or 'regsaude' for "regiões de saúde".}

\item{multi}{Indicator multiplier. Defaults to RIPSA recommendation.}
}
\value{
A \code{data.frame} containing the calculated indicator for the aggregation level.
}
\description{
Follows the RIPSA 2012 card \url{http://fichas.ripsa.org.br/2012/c-4}
}
\examples{
c.4 <- indi_c.4(conn, 2010, "mun")
}
