% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi_c.7.R
\name{indi_c.7}
\alias{indi_c.7}
\title{Proporção de óbitos por infecção respiratória aguda em menores de 5 anos de idade - C.7}
\usage{
indi_c.7(conn, ano, multi = 100)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year.}

\item{multi}{Indicator multiplier. Defaults to RIPSA recommendation.}
}
\value{
A \code{data.frame} containing the municipalities IBGE codes (\code{cod_mun}) and the calculated indicator.
}
\description{
Follows the RIPSA 2012 card \url{http://fichas.ripsa.org.br/2012/c-7}
}
\examples{
c.7 <- indi_c.7(conn, 2010)
}
