% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi_c.9.R
\name{indi_c.9}
\alias{indi_c.9}
\title{Taxa de mortalidade específica por causas externas - C.9}
\usage{
indi_c.9(conn, ano, multi = 1e+05)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year.}

\item{multi}{Indicator multiplier. Defaults to RIPSA recommendation.}
}
\value{
A \code{data.frame} containing the municipalities IBGE codes (\code{cod_mun}) and the calculated indicator.
}
\description{
Follows the RIPSA 2012 card \url{http://fichas.ripsa.org.br/2012/c-9}
}
\examples{
c.9 <- indi_c.9(conn, 2010)
}
