% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sinasc.R
\name{get_sinasc}
\alias{get_sinasc}
\title{Get live births data for municipalities}
\usage{
get_sinasc(conn, agr, ano = NULL, lucene_query = NULL)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year of birth.}

\item{lucene_query}{string. A more complex lucene query can be presented here. The string must be informed using sigle quotes. If used, all other filters will be ignored.}

\item{agg}{string. Agregation level. 'mun' for municipalities, 'uf' for "unidades federativas" or 'regsaude' for "regiões de saúde".}
}
\value{
A \code{data.frame} containing number of live births (\code{sinasc}) for the aggregation level.
}
\description{
This function retrieves live births data from PCDaS ElasticSearch cluster for a specified year.
}
\details{
This function uses raw data from the Sistema de Informações de Nascidos Vivos (SINASC) available at the PCDaS ElasticSearch cluster. A documentation about this data can be found at \url{https://bigdata-metadados.icict.fiocruz.br/dataset/sistema-de-informacoes-de-nascidos-vivos-sinasc}.
}
\examples{
sinasc <- get_sinasc_mun(conn = conn, ano = 2010, agr = "mun")
}
