% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi_c.1.2.R
\name{indi_c.1.2}
\alias{indi_c.1.2}
\title{Taxa de mortalidade neonatal tardia - C.1.2}
\usage{
indi_c.1.2(conn, ano, agr, multi = 1000)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year.}

\item{multi}{Indicator multiplier. Defaults to RIPSA recommendation.}

\item{agg}{string. Aggregation level. 'mun' for municipalities, 'uf' for "unidades federativas" or 'regsaude' for "regiões de saúde".}
}
\value{
A \code{data.frame} containing the calculated indicator for the aggregation level.
}
\description{
Follows the RIPSA 2012 card \url{http://fichas.ripsa.org.br/2012/c-1-2}
}
\examples{
c.1.2 <- indi_c.1.2(conn = conn, ano = 2010, agr = "mun")
}
