% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi_c.12.R
\name{indi_c.12}
\alias{indi_c.12}
\title{Taxa de mortalidade específica por diabete melito - C.12}
\usage{
indi_c.12(conn, ano, agr, multi = 1e+05)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year.}

\item{multi}{Indicator multiplier. Defaults to RIPSA recommendation.}

\item{agg}{string. Aggregation level. 'mun' for municipalities, 'uf' for "unidades federativas" or 'regsaude' for "regiões de saúde".}
}
\value{
A \code{data.frame} containing the calculated indicator for the aggregation level.
}
\description{
Follows the RIPSA 2012 card \url{http://fichas.ripsa.org.br/2012/c-12}
}
\examples{
c.12 <- indi_c.12(conn, 2010, "mun")
}
