% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indi_c.16.R
\name{indi_c.16}
\alias{indi_c.16}
\title{Taxa de mortalidade na infância - C.16}
\usage{
indi_c.16(conn, ano, agr, multi = 1000)
}
\arguments{
\item{conn}{Connection object created with \code{\link{pcdas_connect}}.}

\item{ano}{numeric. Year.}

\item{agr}{string. Aggregation level. 'mun' for municipalities, 'uf' for "unidades federativas" or 'regsaude' for "regiões de saúde".}

\item{multi}{Indicator multiplier. Defaults to RIPSA recommendation.}
}
\value{
A \code{data.frame} containing the calculated indicator for the aggregation level.
}
\description{
Follows the RIPSA 2012 card \url{http://fichas.ripsa.org.br/2012/c-16}
}
\examples{
c.16 <- indi_c.16(conn = conn, ano = 2010, agr = "mun")
}
