% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcdas_connect.R
\name{pcdas_connect}
\alias{pcdas_connect}
\title{Creates a connection to the PCDaS ElasticSearch cluster}
\usage{
pcdas_connect(host = "dados-pcdas.icict.fiocruz.br", port = 443,
  transport_schema = "https", user, pwd = getPass())
}
\arguments{
\item{host}{ElasticSearch cluster address from PCDaS.}

\item{port}{Port number.}

\item{transport_schema}{Only https connection is supported.}

\item{user}{Username at PCDaS}

\item{pwd}{Password at PCDaS, retrieved using a interactive dialog.}
}
\value{
A connection instruction to the PCDaS ElasticSearch cluster.
}
\description{
This function creates a connection to the ElasticSearch server from the Plataforma de Ciência de Dados aplicada à Saúde (PCDaS).
A username and password from PCDaS are needed in order to use this package and can be freely obtained at \url{https://bigdata.icict.fiocruz.br}.
}
\examples{
pcdas_connect(user = "your_username", pwd = "your_password")
}
